新APIに更新

+++ Assets/RaceGame/Scripts/Race/RaceOnlyTester.cs
@@ -11,7 +11,7 @@ namespace RaceGame.Race
     public class RaceOnlyTester : MonoBehaviour
     {
         [Inject] private IGameSetting _gameSetting;
-        [SerializeField] private int localPlayerID = 1370;
+        [SerializeField] private long localPlayerID = 1370;
 
         private void Start()
         {
